Format + remove deprecated files
